feat: drew the owl it seems
Implemented hex printing from a given file while accounting for various
formatting features such as number of columns per line and number of
bytes in each group per line. This behavior mimics xxd, and these
behaviors are decided via cli flags.

Additionally, I added the ability for the user to specify a number of
bytes to be read as well as a seek or offset value from which to start
reading in the provided input. These are also able to be configured via
cli flags at runtime.