feat: colorized output
Added colorized output as an option using a cli flag via the colored
crate. I also fixed an issue with printing where the string portion of
the last line wouldn't line up with the other lines if the length of the
input was not evenly divisible by the column size. I fixed this by
manually calculating an offset and adding padding if necessary.