(refactor): cli, input
Refactored the monolithic main function and moved functionality into two
additional file for cli parsing and retrieving input. Additionally, I
fixed a bug in the seek address tracking logic. Previously, when using
seek the offset address in the left column would always start at 0. To
fix this, I set the seek CLI flag as an Option, which I then parse and
pass to the input function. Additionally, this address value is tracked
in the parsing and output logic to ensure that accurate offsets are
displayed.