(fix): read_input
There was a bug in read input due to neglecting to validate seek and
length inputs. There were no checks in place to ensure that the user
wouldn't read past the end of the file. I fixed this by creating two
functions in input: read_from_file and read_from_stdin. read_input now
matches on filename to determine whether one is given and calls the
appropriate function. For reading from a file, I added checks to ensure
that seek doesn't start after the end of the file as well as updating
the number of bytes to read of seek + len > file_size.